eirp distribution via docker command line
