+++ src/harness/testcases/cu_pass/dpa_calculator/features/steps/dpa_neighborhood/cbsd_creation/ap_heights.py
@@ -9,7 +9,7 @@ from cu_pass.dpa_calculator.cbsds_creator.cbsd_height_distributor.height_distrib
     fractional_distribution_to_height_distribution, HeightDistribution
 from testcases.cu_pass.dpa_calculator.features.steps.dpa_neighborhood.cbsd_creation.common_steps.cbsd_creation import \
     ContextCbsdCreation
-from testcases.cu_pass.dpa_calculator.features.steps.dpa_neighborhood.environment.parsers.parse_fractional_distribution import \
+from cu_pass.dpa_calculator.helpers.list_distributor.fractional_distribution.support.fractional_distribution_parser import \
     DISTRIBUTION_REGEX_UNIFORM, parse_fractional_distribution
 
 use_step_matcher("parse")
