+++ src/harness/testcases/cu_pass/dpa_calculator/features/steps/dpa_neighborhood/docker/command_line_options.py
@@ -41,7 +41,8 @@ def step_impl(context: ContextDocker, cbsd_category: CbsdCategories, distributio
 @then("\"{expected_log_portion}\" should be in the output log")
 def step_impl(context: ContextDocker, expected_log_portion: str):
     output_content = get_uploaded_log_content(bucket_name=context.s3_bucket)
-    assert re.search(expected_log_portion, output_content)
+    expected_log_portion_sanitized = expected_log_portion.replace('[', '\[').replace(']', '\]')
+    assert re.search(expected_log_portion_sanitized, output_content)
 
 
 @then("the results should{not_included}include UE results")
